Implemented a load of changes to the DB
Please review these changes and give feedback, it's a good starting place so I can begin writing the code for the backend.

I'm including the model file that can be opened with MySQL workbench, as well as the creation script and a PNG of the ER-Diagram.

This database is LIVE on the development server, the details are in the group me.