Changes to QueryBuilder, Session, some debugging added
Primary change being the implementation of a proper session.
Properties:
A user can only have one session tied to their userid, if there exists an old one when signing in it will be deleted.
Sessions are periodically checked for expiry, the session length is declared as 60 minutes in config/global.php